+++ R/build_workshops.R
@@ -13,30 +13,30 @@
 
 build_workshops <- function(path = ".", download = FALSE, id = NULL, Rmdfiles = NULL, update_template = FALSE, verbose = TRUE) {
 
+  rx <- "^workshop[0-9]{2}-[ef][nr].[Rr]md$"
+
   if (download) {
     if (is.null(id)) stop("`id` must be specified!")
     if (!dir.exists(path)) dir.create(path)
     path2 <- tempfile(tmpdir = path)
     download_workshop_indiv(id, path2, verbose)
-    Rmdfiles <- find_f(path2, "^workshop[0-9]{2}-[ef][nr].[Rr]md$")
+    Rmdfiles <- find_f(path2, rx)
   } else {
-    if (is.null(Rmdfiles)) 
-      Rmdfiles <- find_f(path, "^workshop[0-9]{2}-[ef][nr].[Rr]md$")
+    if (is.null(Rmdfiles))
+      Rmdfiles <- find_f(path, rx)
   }
-  if (!length(Rmdfiles)) stop("No source file found")
+  if (!length(Rmdfiles)) stop("No source file found!")
 
   install_workshops_pkgs(find_f(path, "^pkgs.yaml$"), verbose = verbose)
 
-  # check/update template files 
-  t_files <- find_f(, "^qcbsR.*[msj][ls]$") 
-  path2 <- gsub("/workshop[0-9]{2}-[ef][nr].[Rr]md$", "", Rmdfiles)
+  # check/update template files
+  t_files <- find_f(path, "^qcbsR.*[msj][ls]$")
+  path2 <- gsub(paste0("/", rx), "", Rmdfiles)
   if (!all(template_files(path2) %in% t_files) | update_template)
     update_template(path, verbose = verbose)
 
-
   render_workshops(Rmdfiles, verbose = verbose)
-  success_msg("Workshop", gsub(".*/", "", Rmdfiles),
-      " successfully rendered.\n")
+  success_msg(gsub(".*/", "", Rmdfiles), " successfully rendered.\n")
 
   invisible(NULL)
 }